requirements.txt
beautifulsoup4==4.12.3
google-generativeai==0.7.2
python-dotenv==1.0.1
python-telegram-bot[job-queue]==21.4
requests==2.32.3
schedule==1.2.2
prometheus-client==0.20.0
aiohttp==3.9.5
tenacity==8.5.0
click==8.1.7
fastapi==0.111.0
uvicorn[standard]==0.30.1
jinja2==3.1.4
mistralai==0.4.2
sqlalchemy
psycopg2-binary
alembic
pytest
APScheduler
pytz